==> R/1_sim_generate_landscapes.qmd <==
---
title: "The 4<sup>th</sup> Dimension in Animal Movement"
subtitle: "Simulation Study: Landscapes"
author:
  - name: Johannes Signer
    email: jsigner@uni-goettingen.de
  - name: Cédric Scherer
    email: viz@cedricscherer.com
  - name: Stephanie Kramer-Schadt
    email: kramer@izw-berlin.de
date: "`r Sys.Date()`"
format:
    html: 
      theme: flatly
      embed-resources: true
      toc: true
      backgroundcolor: '#ffffff'
knitr:
  opts_chunk:    
    dev: "ragg_png"
    retina: 1
    dpi: 300
execute:
  freeze: auto
  echo: true
  warning: false
  message: false
  cache: true
  fig-width: 7.8
  fig-height: 7.8
editor: source
editor_options:
   chunk_output_type: inline
---

## Preparation

```{r}
#| label: setup
#| cache: false
## packages
library(readr)
library(purrr)
library(NLMR)
library(raster)
library(here)

## path + folder for generated landscapes
dir_lscps <- here("data", "simulated", "landscapes")
if (!dir.exists(dir_lscps)) dir.create(dir_lscps, recursive = TRUE)
```


## Simulation Environmental Variables 

```{r}
#| label: simulate-landscapes
#| message: false
frac <- c(0.01, 0.1, 1)
r <- 2e3
reps <- 10
set.seed(221)

## elevation (numeric) and habitat (categorical)
if (!file.exists(paste0(dir_lscps, "/", reps, "/cat.rds"))) {
  for (i in 1:reps) {
    dir.create(paste0(dir_lscps, "/", i), showWarnings = FALSE)
    cont <- map(frac, ~ {
      x <- nlm_fbm(r, r, fract_dim = .x)
      crs(x) <- "epsg:3035"
      names(x) <- "var1"
      x
    })
  
    cat <- map(frac, ~ {
      x <- nlm_fbm(r, r, fract_dim = .x) > 0.5
      crs(x) <- "epsg:3035"
      names(x) <- "var2"
      x
    })
  
    saveRDS(cont, paste0(dir_lscps, "/", i, "/cont.rds"))
    saveRDS(cat, paste0(dir_lscps, "/", i, "/cat.rds"))
  }
} else {
  cont <- read_rds(paste0(dir_lscps, "/", reps, "/cont.rds"))
  cat <- read_rds(paste0(dir_lscps, "/", reps, "/cat.rds"))
}

## homing tendency (hr)
if (!file.exists(paste0(dir_lscps, "/", reps, "/d.rds"))) {
  d <- raster(terra::distance(terra::rast(cat[[1]]), terra::vect(cbind(r/2, r/2), crs = "epsg:3035")))
  names(d) <- "d"

  saveRDS(d, paste0(dir_lscps, "/hr.rds"))
} else {
  d <- read_rds(paste0(dir_lscps, "/hr.rds"))
}
```


## Visualize Landscapes

```{r}
#| label: plots
plot(stack(cont))
plot(stack(cat))
plot(d)
```

*** 
  
<details><summary>Session Info</summary>

```{r}
#| label: session-info
devtools::session_info()
```

</details>
